requirements.txt
amqp==5.3.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cramjam==2.10.0
Django==5.2
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
drf-yasg==1.21.10
inflection==0.5.1
kombu==5.5.3
minio==7.2.15
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2==2.9.10
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.9.0
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-snappy==0.7.3
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
winkerberos==0.12.2
